backend/requirements.txt
# Core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-multipart==0.0.12

# Database
sqlalchemy==2.0.36
alembic==1.14.0
asyncpg==0.30.0
psycopg2-binary==2.9.10

# Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.0.1  # Pin to 4.0.1 for compatibility with passlib
python-dotenv==1.0.1

# OAuth2
authlib==1.3.0
httpx==0.27.0
itsdangerous==2.1.2

# Validation
pydantic==2.9.2
pydantic-settings==2.6.0
email-validator==2.2.0

# Utils
python-dateutil==2.9.0

# CORS
fastapi-cors==0.0.6
